Валидация перенесена в отдельный модуль. Изменена валидация комнат и гостей
