+++ js/main.js
@@ -2,7 +2,8 @@ import { getSettings } from './settings.js';
 import { getElements } from './elements.js';
 import { createAdvertisement } from './create-advertisement.js';
 import { generateCardElements } from './generate-card-elements.js';
-import { putInactiveState, validateForm } from './form.js';
+import { putInactiveState } from './form.js';
+import { validateForm } from './validate-form.js';
 
 
 const {adForm, mapFilters} = getElements();
